Move api server to its own package
